Apply prompt size static asserts to method names
